Refactor for clarity and remove dead code
- Remove unused last_kb_val static and a redundant pad_gap call that was
  immediately overwritten by lv_obj_remove_style_all
- Drop unused IMAGE_PROXY and COLOUR_LINK_BG macros
- Collapse duplicated "undo LVGL-typed label" blocks in the keyboard
  callback into a kb_undo_label_text() helper (behavior-identical)
- Fix %d/lv_coord_t format mismatch in touch debug line

+++ src/touch.cpp
@@ -66,7 +66,7 @@ static void touch_read_cb(lv_indev_drv_t *drv, lv_indev_data_t *data) {
         s_touch_active = true;
         if (s_touch_dbg_count < 5) {
             dbg("TOUCH: raw(%d,%d) -> landscape(%d,%d)", tx, ty,
-                data->point.x, data->point.y);
+                (int)data->point.x, (int)data->point.y);
             s_touch_dbg_count++;
         }
     } else {
